moved services to shared folder
