+++ src/client/entities/user-information/__tests__/use-user-information.store.spec.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, vi } from 'vitest';
-import { localStorageService } from '../../../services/local-storage/local-storage.service';
+import { localStorageService } from '../../../shared/services/local-storage/local-storage.service';
 import { UserInformationKey } from '../../../shared/constants/local-storage.contants';
 import { UserInformation } from '../models/user-infomation.model';
 import { ThemeEnum } from '../../../shared/enums/theme.enum';
